Put proxy source into a valid source directory Add dependencies to build.grade Make changes to keep up with changed interfaces
