+++ examples/proxy/build.gradle
@@ -2,6 +2,5 @@ apply plugin: 'java'
 
 dependencies {
     compile project(':msl-core')
+    compile 'io.reactivex:rxjava:latest.release'
 }
-
-jar.baseName = "msl-proxy"
