+++ examples/proxy/src/main/java/com/netflix/msl/ProxyException.java
