+++ examples/proxy/src/main/java/com/netflix/msl/entityauth/FailingEntityAuthenticationFactory.java
