+++ examples/proxy/src/main/java/com/netflix/msl/entityauth/ProxyEntityAuthenticationScheme.java
