+++ examples/proxy/src/main/java/com/netflix/msl/tokens/FailoverTokenFactory.java
@@ -5,26 +5,27 @@ package com.netflix.msl.tokens;
 
 import javax.crypto.SecretKey;
 
+import com.netflix.msl.MslCryptoException;
+import com.netflix.msl.MslEncodingException;
 import com.netflix.msl.MslError;
 import com.netflix.msl.MslException;
 import com.netflix.msl.ProxyMslError;
+import com.netflix.msl.entityauth.EntityAuthenticationData;
+import com.netflix.msl.userauth.ProxyMslUser;
 import com.netflix.msl.util.MslContext;
 
 /**
- * <p>This token factory accepts all tokens and throws an exception if a token
- * must be created or renewed.</p>
+ * <p>This token factory accepts all tokens and renews tokens by echoing them
+ * back.</p>
  * 
  * @author Wesley Miaw <wmiaw@netflix.com>
  */
-public class ProxyTokenFactory implements TokenFactory {
+public class FailoverTokenFactory implements TokenFactory {
     /* (non-Javadoc)
      * @see com.netflix.msl.tokens.TokenFactory#isMasterTokenRevoked(com.netflix.msl.util.MslContext, com.netflix.msl.tokens.MasterToken)
      */
     @Override
     public MslError isMasterTokenRevoked(final MslContext ctx, final MasterToken masterToken) {
-        // TODO This check should be implemented to immediately check if a
-        // master token has been revoked. For now return null indicating the
-        // master token is acceptable.
         return null;
     }
 
@@ -33,16 +34,14 @@ public class ProxyTokenFactory implements TokenFactory {
      */
     @Override
     public MslError acceptNonReplayableId(final MslContext ctx, final MasterToken masterToken, final long nonReplayableId) throws MslException {
-        // TODO This check should be implemented somehow. For now throw an
-        // exception to trigger processing by the proxied MSL service.
-        throw new MslException(ProxyMslError.NONREPLAYABLE_ID_CHECK_REQUIRED);
+        return null;
     }
 
     /* (non-Javadoc)
-     * @see com.netflix.msl.tokens.TokenFactory#createMasterToken(com.netflix.msl.util.MslContext, java.lang.String, javax.crypto.SecretKey, javax.crypto.SecretKey)
+     * @see com.netflix.msl.tokens.TokenFactory#createMasterToken(com.netflix.msl.util.MslContext, com.netflix.msl.entityauth.EntityAuthenticationData, javax.crypto.SecretKey, javax.crypto.SecretKey)
      */
     @Override
-    public MasterToken createMasterToken(final MslContext ctx, final String identity, final SecretKey encryptionKey, final SecretKey hmacKey) throws MslException {
+    public MasterToken createMasterToken(MslContext ctx, EntityAuthenticationData entityAuthData, SecretKey encryptionKey, SecretKey hmacKey) throws MslEncodingException, MslCryptoException, MslException {
         // This method should not get called. If it does then throw an
         // exception to trigger processing by the proxied MSL service.
         throw new MslException(ProxyMslError.MASTERTOKEN_CREATION_REQUIRED);
@@ -53,9 +52,6 @@ public class ProxyTokenFactory implements TokenFactory {
      */
     @Override
     public MslError isMasterTokenRenewable(final MslContext ctx, final MasterToken masterToken) {
-        // Assume the master token will be renewed if it needs to be. The
-        // downside of not checking right now is that we may reject the message
-        // after doing application-level work.
         return null;
     }
 
@@ -64,7 +60,7 @@ public class ProxyTokenFactory implements TokenFactory {
      */
     @Override
     public MasterToken renewMasterToken(final MslContext ctx, final MasterToken masterToken, final SecretKey encryptionKey, final SecretKey hmacKey) throws MslException {
-        throw new MslException(ProxyMslError.MASTERTOKEN_RENEWAL_REQUIRED);
+        return masterToken;
     }
 
     /* (non-Javadoc)
@@ -72,9 +68,6 @@ public class ProxyTokenFactory implements TokenFactory {
      */
     @Override
     public MslError isUserIdTokenRevoked(final MslContext ctx, final MasterToken masterToken, final UserIdToken userIdToken) {
-        // TODO This check should be implemented to immediately check if a
-        // user ID token has been revoked. For now return null indicating the
-        // user ID token is acceptable.
         return null;
     }
 
@@ -91,7 +84,7 @@ public class ProxyTokenFactory implements TokenFactory {
      */
     @Override
     public UserIdToken renewUserIdToken(final MslContext ctx, final UserIdToken userIdToken, final MasterToken masterToken) throws MslException {
-        throw new MslException(ProxyMslError.USERIDTOKEN_RENEWAL_REQUIRED);
+        return userIdToken;
     }
 
     /* (non-Javadoc)
@@ -99,6 +92,6 @@ public class ProxyTokenFactory implements TokenFactory {
      */
     @Override
     public MslUser createUser(final MslContext ctx, final String userdata) {
-        return new NetflixMslUser(userdata);
+        return new ProxyMslUser(userdata);
     }
 }
