+++ examples/proxy/src/main/java/com/netflix/msl/userauth/FailingUserAuthenticationFactory.java
