+++ examples/proxy/src/main/java/com/netflix/msl/userauth/ProxyUserAuthenticationScheme.java
