+++ examples/proxy/src/main/java/com/netflix/msl/util/FailoverMslContext.java
